(PC-32694)[PRO] refactor: reorder CollectiveOfferSummary components and pages
	- put src/components/CollectiveOfferSummary inside src/pages/CollectiveOffer/CollectiveOfferSummary/
	- put CollectiveOfferSummaryCreationScreen next to CollectiveOfferSummaryCreation as they soon will be merged together
