+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/CollectiveOfferSummaryCreation/__specs__/CollectiveOfferSummaryCreationScreen.spec.tsx
@@ -1,4 +1,5 @@
 import { screen } from '@testing-library/react'
+import { expect } from 'vitest'
 
 import {
   getCollectiveOfferFactory,
