+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/components/CollectiveOfferSummary/__specs__/CollectiveOfferSummary.spec.tsx
