+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/components/CollectiveOfferSummary/components/CollectiveOfferLocationSection.tsx
