+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/components/CollectiveOfferSummary/components/__specs__/CollectiveOfferLocationSection.spec.tsx
