+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/components/CollectiveOfferSummary/components/__specs__/CollectiveOfferVisibilitySection.spec.tsx
