+++ pro/src/pages/CollectiveOffer/CollectiveOfferSummary/components/CollectiveOfferSummary/components/utils/__specs__/formatDuration.spec.tsx
